+++ CHANGELOG.md
@@ -16,6 +16,7 @@ All notable changes to this project will be documented in this file, following t
 - Add `element_granularity` field to component expressions
 - Add I/HM basic restraints example
 - Add primitives: ellipse, ellipsoid, box, arrow
+- Add basic support for volumetric data
 
 
 ## [v1.1.0] - 2024-12-09